Add Pages::GA, Pages::Committee, and Pages::Caucus
This adds subclasses of Page for GA, Committee, and Caucus. Through this
heirarchy it looks like we'll be able to inherit the functionality we
give to pages.

For now, routes exist so that each type of page exists in its own
namespace -- under /pages, /generalassembly, /committee, /caucus for
now.

I'm not 100% sure this is the correct way to go about this -- but it
seems a heck of a lot more correct than whatever we were trying before.

My thinking for the customization of each of the types of page is that
we will, instead of storing a string of "content" we could store a JSON
object which corresponds to different parts of the page. Editing a part
with Mercury editor would then update the JSON for that page.